requirements.txt
robyn==0.37.0
uvloop==0.17.0
markdown2==2.4.10
python-frontmatter==1.0.0